python: re-organize
