+++ python/CMakeLists.txt
@@ -1,4 +1,11 @@
-set(${PROJECT_NAME}_PYTHON_FILES robots_loader.py __main__.py __init__.py)
+set(${PROJECT_NAME}_PYTHON_FILES
+    human.py
+    panda.py
+    robots_loader.py
+    talos.py
+    utils.py
+    __main__.py
+    __init__.py)
 
 foreach(python ${${PROJECT_NAME}_PYTHON_FILES})
   python_install_on_site(${PY_NAME} ${python})
